refactor: improved organisation and reuseability of scripts
